% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_list_1}
\alias{dictionary_list_1}
\title{Data-driven dictionary list 1}
\format{
An object of class \code{list} of length 233.
}
\usage{
dictionary_list_1
}
\description{
Contains a dictionary using wildcard values and compound words from \code{SpecialEduTech}
}
\examples{
dictionary_list_1

}
\keyword{datasets}
