% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{examine_gamma_terms}
\alias{examine_gamma_terms}
\title{Examine highest document-topic probabilities}
\usage{
examine_gamma_terms(gamma_td, top_terms, ...)
}
\arguments{
\item{gamma_td}{A tidy data frame that includes document-topic probabilities.
(probabilities of each topic per each document).}

\item{top_terms}{A tidy data frame that contains top terms with highest
document-topic probabilities within each topic.}

\item{...}{Further arguments passed to \code{group_by}.}
}
\description{
Examine highest document-topic probabilities.
}
\examples{
\dontrun{
library(tidytext)

# Tidy the document-topic combinations
gamma_td <- tidy(stm_15, matrix = "gamma",
                document_names = rownames(dfm))
gamma_td \%>\% examine_gamma_terms(top_terms = top_terms_selected)
}

}
