% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{examine_top_terms}
\alias{examine_top_terms}
\title{Examine highest word probabilities for each topic}
\usage{
examine_top_terms(beta_td, top_n, ...)
}
\arguments{
\item{beta_td}{A tidy data frame that includes term-topic probabilities
(probabilities of each word per each topic).}

\item{top_n}{A number of top n terms with highest term-topic probabilities in each document.}

\item{...}{Further arguments passed to \code{group_by}.}
}
\description{
Examine highest document-topic probabilities.
}
\examples{
\dontrun{
library(tidytext)

# Tidy the word-topic combinations
beta_td <- tidy(stm_15, document_names = rownames(dfm))
beta_td \%>\% examine_top_terms(top_n = 5)
}

}
