% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_topic_probability}
\alias{plot_topic_probability}
\title{Visualize a plot for document-topic probabilities}
\usage{
plot_topic_probability(gamma_terms, top_n, topic_names = NULL)
}
\arguments{
\item{gamma_terms}{A tidy data frame that includes document-topic probabilities.
(probabilities of each topic per each document).}

\item{top_n}{A number of top n terms with highest document-topic probabilities.}

\item{topic_names}{Topic names}
}
\description{
Visualize a plot for document-topic probabilities.
}
\examples{
\dontrun{
library(tidytext)

# Tidy the document-topic combinations
top_terms_selected <- beta_td \%>\% examine_top_terms(top_n = 5)
gamma_td <- tidy(stm_15, matrix = "gamma",
                document_names = rownames(dfm))
gamma_terms <- gamma_td \%>\% examine_gamma_terms(top_terms = top_terms_selected)
gamma_terms \%>\% plot_topic_probability(top_n = 15)
}

}
