% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_frequency}
\alias{plot_word_frequency}
\title{Plot word frequency results.}
\usage{
plot_word_frequency(data, n = 20, ...)
}
\arguments{
\item{data}{A document-feature matrix (dfm) object through the quanteda package.}

\item{n}{The number of top n features (terms or words).}

\item{...}{Further arguments passed to \code{quanteda.textstats::textstat_frequency}.}
}
\description{
Plot the frequently observed top n terms.
}
\examples{
\dontrun{
dfm <- processed_tokens \%>\% construct_dfm()

dfm \%>\% plot_word_frequency(n = 20)
}

}
