% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{preprocess_texts}
\alias{preprocess_texts}
\title{Preprocess text data}
\usage{
preprocess_texts(data, text_field = "united_texts", ...)
}
\arguments{
\item{data}{A data frame that contains text as data.}

\item{text_field}{A name of column that contains text data in a data frame.}

\item{...}{Further arguments passed to \code{corpus}.}
}
\description{
Preprocess text data by conducting the following functions:
construct a corpus; segment texts in a corpus into tokens; preprocess tokens;
convert the features of tokens to lowercase;
remove stopwords; specify the minimum length in characters for tokens (at least 2).
}
\examples{
\dontrun{
library(dplyr)

processed_tokens <- SpecialEduTech \%>\% preprocess_texts(text_field = "abstract")

}

}
