% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpecialEduTech}
\alias{SpecialEduTech}
\title{Special education technology bibliographic data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 490 rows and 6 columns.
}
\usage{
SpecialEduTech
}
\description{
Contains bibliographic data of journal articles and dissertations on the use of technology in teaching mathematics to students with disabilities.
}
\examples{
SpecialEduTech

}
\keyword{datasets}
