% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_list}
\alias{stopwords_list}
\title{Data-driven stopword list}
\format{
An object of class \code{character} of length 117.
}
\usage{
stopwords_list
}
\description{
Contains stopword list that was created based on inverse document frequency from \code{SpecialEduTech}
}
\examples{
stopwords_list

}
\keyword{datasets}
