% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textminingR-package.R
\docType{package}
\name{textminingR-package}
\alias{textminingR}
\alias{textminingR-package}
\title{textminingR: Text Mining Workflow Tools}
\description{
Text mining analysis tools using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mshin77.github.io/textminingR}
}

}
\author{
\strong{Maintainer}: Mikyung Shin \email{shin.mikyung@gmail.com} (\href{https://orcid.org/0000-0001-7907-9193}{ORCID})

}
\keyword{internal}
