% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textminingR.app.R
\name{textminingR.app}
\alias{textminingR.app}
\title{Launch and browser the textminingR app}
\usage{
textminingR.app()
}
\description{
Launch and browser the textminingR app.
}
\examples{
\dontrun{
library(textminingR)
textminingR.app()
}
}
